requirements.txt
# CLIP模型 如果报错可下载包 https://github.com/openai/CLIP，手动安装 pip install .
git+https://github.com/openai/CLIP.git

# 基础依赖
numpy==1.23.5
Pillow==10.0.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86

# 设备控制
pure-python-adb==0.3.0.dev0

# AI模型相关
torch==2.0.1
openai==1.3.5
httpx==0.25.2
scikit-learn==1.3.0

# OCR相关
paddlepaddle==3.0.0b1
paddleocr==2.7.0.3

# 工具依赖
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# 开发工具
pytest==7.4.0
black==23.7.0
isort==5.12.0


# # 对于Apple Silicon Mac
# pip install paddlepaddle==3.0.0b1 -f https://www.paddlepaddle.org.cn/whl/mac/avx/stable.html
